refactor(sample): extract SAMPLES and ThemePair into SampleData.kt
Moves the large SAMPLES list and ThemePair data class out of SampleScreen.kt
into a dedicated SampleData.kt file, keeping SampleScreen.kt focused on
composable structure and navigation logic.
